requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
db-dtypes==1.3.0
google-api-core==2.21.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
load-dotenv==0.1.0
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pandas-gbq==0.24.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
google-cloud-secret-manager